requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psutil==5.9.6
python-dotenv==1.0.0
google-generativeai==0.3.1
vertexai==1.38.0
aiohttp==3.9.1
pydantic==2.4.2
python-multipart==0.0.6 